chore: add SQL scripts for database setup
This scripts can be used to create and populate the local database with
an extension like SQLTools. I prefer this during development rather than
running the createTables.js script each time I want to update the
database.

Even so, the createdb.js script now creates and also populates the
database, which could still be useful to seed the production database.